chore: Organize WOW services and tests
